+++ crates/store/src/db/tests.rs
@@ -22,6 +22,7 @@ use miden_objects::account::{
     AccountStorageMode,
     AccountType,
     AccountVaultDelta,
+    PublicKeyCommitment,
     StorageSlot,
 };
 use miden_objects::asset::{Asset, FungibleAsset};
@@ -32,13 +33,13 @@ use miden_objects::block::{
     BlockNoteTree,
     BlockNumber,
 };
-use miden_objects::crypto::dsa::rpo_falcon512::PublicKey;
 use miden_objects::crypto::merkle::SparseMerklePath;
 use miden_objects::crypto::rand::RpoRandomCoin;
 use miden_objects::note::{
     Note,
     NoteDetails,
     NoteExecutionHint,
+    NoteHeader,
     NoteId,
     NoteMetadata,
     NoteTag,
@@ -52,7 +53,13 @@ use miden_objects::testing::account_id::{
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE,
     ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE_2,
 };
-use miden_objects::transaction::{OrderedTransactionHeaders, TransactionHeader, TransactionId};
+use miden_objects::transaction::{
+    InputNoteCommitment,
+    InputNotes,
+    OrderedTransactionHeaders,
+    TransactionHeader,
+    TransactionId,
+};
 use miden_objects::{EMPTY_WORD, Felt, FieldElement, Word, ZERO};
 use pretty_assertions::assert_eq;
 use rand::Rng;
@@ -1098,21 +1105,25 @@ fn notes() {
     let block_num_1 = 1.into();
     create_block(conn, block_num_1);
 
+    let block_range = BlockNumber::from(0)..=BlockNumber::from(1);
+
     // test empty table
-    let res =
-        queries::select_notes_since_block_by_tag_and_sender(conn, BlockNumber::from(0), &[], &[])
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[], block_range.clone())
             .unwrap();
 
     assert!(res.is_empty());
+    assert_eq!(last_included_block, 1.into());
 
-    let res = queries::select_notes_since_block_by_tag_and_sender(
+    let (res, last_included_block) = queries::select_notes_since_block_by_tag_and_sender(
         conn,
-        BlockNumber::from(0),
         &[],
         &[1, 2, 3],
+        block_range.clone(),
     )
     .unwrap();
     assert!(res.is_empty());
+    assert_eq!(last_included_block, 1.into());
 
     let sender = AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap();
 
@@ -1151,25 +1162,27 @@ fn notes() {
     queries::insert_notes(conn, &[(note.clone(), None)]).unwrap();
 
     // test empty tags
-    let res =
-        queries::select_notes_since_block_by_tag_and_sender(conn, BlockNumber::from(0), &[], &[])
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[], block_range.clone())
             .unwrap();
     assert!(res.is_empty());
+    assert_eq!(last_included_block, 1.into());
+
+    let block_range_1 = 1.into()..=1.into();
 
     // test no updates
-    let res = queries::select_notes_since_block_by_tag_and_sender(conn, block_num_1, &[], &[tag])
-        .unwrap();
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[tag], block_range_1)
+            .unwrap();
     assert!(res.is_empty());
+    assert_eq!(last_included_block, 1.into());
 
     // test match
-    let res = queries::select_notes_since_block_by_tag_and_sender(
-        conn,
-        block_num_1.parent().unwrap(),
-        &[],
-        &[tag],
-    )
-    .unwrap();
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[tag], block_range.clone())
+            .unwrap();
     assert_eq!(res, vec![note.clone().into()]);
+    assert_eq!(last_included_block, 1.into());
 
     let block_num_2 = note.block_num + 1;
     create_block(conn, block_num_2);
@@ -1186,20 +1199,24 @@ fn notes() {
 
     queries::insert_notes(conn, &[(note2.clone(), None)]).unwrap();
 
+    let block_range = 0.into()..=2.into();
+
     // only first note is returned
-    let res = queries::select_notes_since_block_by_tag_and_sender(
-        conn,
-        block_num_1.parent().unwrap(),
-        &[],
-        &[tag],
-    )
-    .unwrap();
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[tag], block_range)
+            .unwrap();
     assert_eq!(res, vec![note.clone().into()]);
+    assert_eq!(last_included_block, 1.into());
+
+    let block_range = 1.into()..=2.into();
 
     // only the second note is returned
-    let res = queries::select_notes_since_block_by_tag_and_sender(conn, block_num_1, &[], &[tag])
-        .unwrap();
+    let (res, last_included_block) =
+        queries::select_notes_since_block_by_tag_and_sender(conn, &[], &[tag], block_range)
+            .unwrap();
     assert_eq!(res, vec![note2.clone().into()]);
+    assert_eq!(last_included_block, 2.into());
+
     // test query notes by id
     let notes = vec![note.clone(), note2];
 
@@ -1389,6 +1406,24 @@ fn mock_block_transaction(account_id: AccountId, num: u64) -> TransactionHeader 
     let input_notes_commitment = Word::try_from([0, 0, num, 0]).unwrap();
     let output_notes_commitment = Word::try_from([0, 0, 0, num]).unwrap();
 
+    let notes = vec![InputNoteCommitment::from(num_to_nullifier(num))];
+    let input_notes = InputNotes::new_unchecked(notes);
+
+    let output_notes = vec![NoteHeader::new(
+        NoteId::new(
+            Word::try_from([num, num, 0, 0]).unwrap(),
+            Word::try_from([0, 0, num, num]).unwrap(),
+        ),
+        NoteMetadata::new(
+            account_id,
+            NoteType::Public,
+            NoteTag::LocalAny(num as u32),
+            NoteExecutionHint::None,
+            Felt::default(),
+        )
+        .unwrap(),
+    )];
+
     TransactionHeader::new_unchecked(
         TransactionId::new(
             initial_state_commitment,
@@ -1399,11 +1434,8 @@ fn mock_block_transaction(account_id: AccountId, num: u64) -> TransactionHeader 
         account_id,
         initial_state_commitment,
         final_account_commitment,
-        vec![num_to_nullifier(num)],
-        vec![NoteId::new(
-            Word::try_from([num, num, 0, 0]).unwrap(),
-            Word::try_from([0, 0, num, num]).unwrap(),
-        )],
+        input_notes,
+        output_notes,
     )
 }
 
@@ -1465,7 +1497,7 @@ fn mock_account_code_and_storage(
         .storage_mode(storage_mode)
         .with_assets(assets)
         .with_component(component)
-        .with_auth_component(AuthRpoFalcon512::new(PublicKey::new(EMPTY_WORD)))
+        .with_auth_component(AuthRpoFalcon512::new(PublicKeyCommitment::from(EMPTY_WORD)))
         .build_existing()
         .unwrap()
 }